requirements.txt
numpy==1.17.4
PyOpenGL==3.1.0